pyproject.toml
[project]
name = "auto-sign"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
